posts/china_ai/requirements.txt
modal
notebook
pillow